requirements.txt
backports.ssl-match-hostname==3.4.0.2
requests==2.20.0
six==1.9.0
websocket-client==0.29.0
wsgiref==0.1.2